refactor: sepxr writer
